+++ src/api/UploadAPI.ts
@@ -3,7 +3,8 @@ import base, {DirectUploadInterface} from './base'
 import UploadClient from '../UploadClient'
 import {RequestOptions, RequestResponse} from './request'
 import {FileData, Settings} from '../types'
-import info, {Uuid} from './info'
+import info from './info'
+import {Token, Uuid} from './types'
 import {InfoResponse} from './info'
 import fromUrl, {FromUrlResponse, Url} from './fromUrl'
 import fromUrlStatus, {FromUrlStatusResponse} from './fromUrlStatus'
@@ -17,7 +18,7 @@ export interface UploadAPIInterface {
 
   fromUrl(sourceUrl: Url, settings?: Settings): Promise<FromUrlResponse>
 
-  fromUrlStatus(token: string, settings?: Settings): Promise<FromUrlStatusResponse>
+  fromUrlStatus(token: Token, settings?: Settings): Promise<FromUrlStatusResponse>
 }
 
 class UploadAPI implements UploadAPIInterface {
@@ -46,7 +47,7 @@ class UploadAPI implements UploadAPIInterface {
     return base(data, this.getResultSettings(settings))
   }
 
-  info(uuid: string, settings: Settings = {}): Promise<InfoResponse> {
+  info(uuid: Uuid, settings: Settings = {}): Promise<InfoResponse> {
     return info(uuid, this.getResultSettings(settings))
   }
 
@@ -54,7 +55,7 @@ class UploadAPI implements UploadAPIInterface {
     return fromUrl(sourceUrl, this.getResultSettings(settings))
   }
 
-  fromUrlStatus(token: string, settings: Settings = {}): Promise<FromUrlStatusResponse> {
+  fromUrlStatus(token: Token, settings: Settings = {}): Promise<FromUrlStatusResponse> {
     return fromUrlStatus(token,this.getResultSettings(settings))
   }
 }