+++ src/index.ts
@@ -1,3 +1,12 @@
 import UploadClient from './UploadClient'
+export {Settings, FileData, OriginalImageInfo, UploadcareFile} from './types'
+export {Url} from './api/fromUrl'
+export {Uuid} from './api/types'
+export {UploadFromInterface} from './fileFrom/UploadFrom'
+export {UploadClientInterface} from './UploadClient'
+export {UploadAPIInterface} from './api/UploadAPI'
+export {DirectUploadInterface} from './api/base'
+export {RequestOptions, RequestInterface} from './api/request'
+export {FileFrom} from './fileFrom/fileFrom'
 
 export default UploadClient